requirements.txt
fastapi==0.92.0
uvicorn==0.20.0
python-dotenv==1.0.0
psycopg2-binary==2.9.5
pydantic==1.10.4